Add the topic to the callback method before executing the callback
This allows the developer to access the current topic being handled
by the callback without breaking existing implementations of
amplify pub/sub. The example in the readme shows how to access this
new property.

Also update the readme to clarify that multiple topics can be
subscribed in a single subscribe call.