Added null return if there are no subscriptions.
Currently, publish method returns true or false, in the documentation, it says 

"amplify.publish returns a boolean indicating whether any subscriptions returned false. The return value is true if none of the subscriptions returned false, and false otherwise. Note that only one subscription can return false because doing so will prevent additional subscriptions from being invoked."

Quoting: "The return value is true if none of the subscriptions returned false, and false otherwise."

What if there are no subscriptions yet? It will still return true? 

So this is a suggestion to the feature to inform that when there are not subscriptions yet it will return null.